requirements.txt
# for fpgen
selenium
tbselenium
pypcapfile
matplotlib
IPython

# for crawling
beautifulsoup4
pyaml

# for classification
sklearn
pandas
